% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signature_matrices.R
\name{get_signature_matrices}
\alias{get_signature_matrices}
\title{Get signature matrices.}
\usage{
get_signature_matrices(type = "both")
}
\arguments{
\item{type}{a character vector that can be 'both', 'pVal', or 'OR'}
}
\value{
\code{get_signature_matrices} Returns a list of signature matrices currently stored in scMappR_Data. \cr
}
\description{
This function downloads and returns signature matrices from the scMappR_data repo.
}
\examples{
\donttest{
 
signatures <- get_signature_matrices(type = "both")
 
}
 
}
