% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt_example.R
\name{gmt_example}
\alias{gmt_example}
\title{gmt_example}
\format{A list with 5 characters, each containing genes.
\describe{
  \item{Astrocytes_panglao}{ astrocyte markers identified by panglao}
  \item{Schwann_panglao}{ Schwann markers identified by panglao}
  \item{Bergmann glia_panglao}{ Bergmann glia markers identified by panglao}
  \item{Kupffer_panglao}{ Kupffer markers identified by panglao}
  \item{Oligodendrocyte progenitor_panglao}{ Oligodendrocyte progenitor markers identified by panglao}
}}
\usage{
data(gmt_example)
}
\description{
Markers of 5 glial cell-types
}
\details{
A named list containing the cell-type markers of 5 glial cell types. Used for testing cell-type naming functions.
}
\examples{
data(gmt_example)
}
\keyword{datasets}
