% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_mouse_ct_marker_enrich.R
\name{human_mouse_ct_marker_enrich}
\alias{human_mouse_ct_marker_enrich}
\title{Consensus cell-type naming (Fisher's Exact)}
\usage{
human_mouse_ct_marker_enrich(gene_lists, theSpecies = "human",
  cell_marker_path = "~/scMappR/data", naming_preference = -9)
}
\arguments{
\item{gene_lists}{A named list of vectors containing cell-type markers (mouse or human gene-symbols) which will be named as a cell-type via the fish'er sexact test method.}

\item{theSpecies}{The species of the gene symbols: human or mouse.}

\item{cell_marker_path}{If local, path to Cell-Type marker rda files, otherwise, we will try to download datafiles.}

\item{naming_preference}{Either -9 if there is no expected cell-type or one of the categories from get_naming_preference_options(). This is useful if you previously have an idea of which cell-type you were going to enrich.}
}
\value{
\code{human_mouse_ct_marker_enrich} list: The top cell-type for each cell-type markers list, as well as a matrix the top 5 most likely cell-type makers for each gene list (with OR and P-val) \cr
}
\description{
This function completes the Fisher's exact test cell-type naming for all cell-types
}
\details{
Fisher's exact test method of cell-type identification using the Panglao and CellMarker databases. It extracts significant pathways (pFDR < 0.05).
Then, if naming_preference != -9, it will extract the enriched cell-types within the cell-types identified with the naming preferences option.
Generally, this method seems to be biased to cell-types with a greater number of markers.
}
\examples{
\\notrun {

# load in signature matrices
load("~/scMappR/data/Preoptic_region_example.rda")
# data(Preoptic_region_example)
Signature <- POA_Rank_signature
 rowname <- get_gene_symbol(Signature)
 rownames(Signature) <- rowname$rowname
 genes <- rownames(Signature)[1:100]
 
 # Assuming mouse_cell_markers.rda is in you "~/scMappR/data" directory
 gmt1 <- "~/scMappR/data/mouse_cell_markers.rda"
 gmt <- gmt_panglao
 enriched <- human_mouse_ct_marker_enrich(gene_lists = genes, theSpecies = "mouse", cell_marker_path = "", naming_preference = "brain")
 }
}
