% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBMC_example.R
\name{PBMC_example}
\alias{PBMC_example}
\title{PBMC_scMappR}
\format{A list containing three data frames, normalized count data, a signature matrix, and a list of differentially expressed genes.
\describe{
  \item{bulk_normalized}{ A 3231 x 9 matrix where rows are genes, columns are samples, and the matrix is filled with CPM normalized counts.}
  \item{odds_ratio_in}{ A 2336 x 7 matrix where rows are genes, columns are cell-types and matrix is filled with the odds-ratio that a gene is in each cell-type.}
  \item{bulk_DE_cors}{A 59 x 3 matrix of sex-specific genes found between male and female PBMC samples (female biased = upregulated). row and rownames are genes, columns are gene name, FDR adjusted p-value, and log2 fold-change. DEGs were computed with DESeq2 and genes with a log2FC > 1 were kept. }
}}
\usage{
data(PBMC_example)
}
\description{
Data for scMappR_and_pathway_analysis example
}
\details{
A named list called "PBMC_example", containing the count data, signature matrix, and DEGs. The count data and signature matrix are shortened to fit the size of the package and do not reflect the paper.
}
\examples{
data(PBMC_example)
}
\keyword{datasets}
