% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMappR_tissues.R
\name{scMappR_tissues}
\alias{scMappR_tissues}
\title{scMappR_tissues}
\format{A vector of tissue names available for tissue_scMappR_internal or to download and use in scMappR_and_pathway_analysis.
\describe{
  \item{scMappR_tissues}{ A list of 174 tissue names from PanglaoDB. }
}}
\usage{
data(scMappR_tissues)
}
\description{
Tissues available in scMappR as of February 2020.
}
\details{
A vector of tissues available in scMappR as of February 2020.
}
\examples{
data(scMappR_tissues)
}
\keyword{datasets}
