% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_process.R
\name{single_cell_process}
\alias{single_cell_process}
\title{single_cell_process}
\format{A 752 x 236 matrix of class dgCMatrix where rows are genes and columns are cells. Data matrix is filled with counts detected from scRNAseq
\describe{
  \item{TCTCTAACACAGGCCT}{ Barcode of one of the sequenced cells present. Each column is the count from a scRNA-seq dataset reprocessed by PanglaoDB.}
}}
\usage{
data(single_cell_process)
}
\description{
Example data for processing scRNA-seq count data with Seurat.
}
\details{
A dgCMatrix object containing count data for scRNA-seq processing.
}
\examples{
data(single_cell_process)
}
\keyword{datasets}
