% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topgenes_extract.R
\name{topgenes_extract}
\alias{topgenes_extract}
\title{Extract Top Markers}
\usage{
topgenes_extract(generes, padj = 0.05, FC = 1.5, topNum = 30)
}
\arguments{
\item{generes}{A list of cell-tpe markers with fold-changes and P-vlaues (FindMarkers output in Seurat)}

\item{padj}{The p-value (FDR) cutoff}

\item{FC}{The fold-change cutoff}

\item{topNum}{The number of genes to extract}
}
\value{
\code{topgenes_extract} Returns a numeric vector. \cr
}
\description{
Internal -- Extracts strongect cell-type markers from a Seurat object
}
\details{
Internal, this function runs through a list of outputs from FindMarkers objects in Seurat
at will take genes past a padj and FC threshold. Then it extracts the topNum number of genes
if you have not used the FindMarkers function, then a list of summary statistics with 
fld change designated by avg_logFC and p-val by p_val_adj
}
\examples{
 
 # load generes object
 data(Preoptic_Area)
 topGenes <- topgenes_extract(POA_example$POA_generes)


}
