% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_method_pathway_enrichment.R
\name{two_method_pathway_enrichment}
\alias{two_method_pathway_enrichment}
\title{two_method_pathway_enrichment}
\usage{
two_method_pathway_enrichment(
  DEGs,
  theSpecies,
  scMappR_vals,
  background_genes = NULL,
  output_directory = "output",
  plot_names = "reweighted",
  number_genes = -9,
  newGprofiler = FALSE,
  toSave = FALSE,
  path = NULL
)
}
\arguments{
\item{DEGs}{Differentially expressed genes (gene_name, padj, log2fc).}

\item{theSpecies}{Human, mouse, or a charcter that is compatible with gProfileR.}

\item{scMappR_vals}{cell weighted Fold-changes of differentially expressed genes.}

\item{background_genes}{A list of background genes to test against. NULL assumes all genes in gprofiler gene set databases.}

\item{output_directory}{Path to the directory where files will be saved.}

\item{plot_names}{Names of output.}

\item{number_genes}{Number of genes to if there are many, many DEGs.}

\item{newGprofiler}{Whether to use gProfileR or gprofiler2 (T/F).}

\item{toSave}{Allow scMappR to write files in the current directory (T/F).}

\item{path}{If toSave == TRUE, path to the directory where files will be saved.}
}
\value{
List with the following elements:
\item{rank_increase}{A list containing the degree of rank change between bulk DE genes and cwFold-changes. Pathway enrichment and tf enrichment of these reranked genes.}
\item{non_rank_increase}{list of DFs containing the pathway and TF enrichment of cwFold-changes.}
}
\description{
Pathway analysis of each cell-type based on cell-type specificity and rank improvement by scMappR.
}
\details{
This function re-ranks cwFoldChanges based on their absolute ct specificity scores (per-celltype) as well as their rank increase in cell-type specificity before completing an ordered pathway analysis. In the second method, only genes with a rank increase in cell-type specificity were included
}
\examples{
\donttest{

data(POA_example)
POA_generes <- POA_example$POA_generes
POA_OR_signature <- POA_example$POA_OR_signature
POA_Rank_signature <- POA_example$POA_Rank_signature
Signature <- POA_Rank_signature
rowname <- get_gene_symbol(Signature)
rownames(Signature) <- rowname$rowname
genes <- rownames(Signature)[1:100]

enriched <- tissue_by_celltype_enrichment(gene_list = genes, 
species = "mouse",p_thresh = 0.05, isect_size = 3)

}
 
}
