% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute_and_contextualize.R
\name{deconvolute_and_contextualize}
\alias{deconvolute_and_contextualize}
\title{Generate scMappR Transformed Values (STVs)}
\usage{
deconvolute_and_contextualize(count_file, signature_matrix, DEG_list,
  case_grep, control_grep, max_proportion_change = -9, print_plots = T,
  plot_names = "scMappR", theSpecies = "human", make_scale = F,
  FC_coef = T, sig_matrix_size = 3000, sig_distort = 1,
  drop_unkown_celltype = TRUE, toSave = FALSE)
}
\arguments{
\item{count_file}{Normalized RNA-seq count matrix where rows are gene symbols and columns are individuals. Either the object tself of the path of a TSV file}

\item{signature_matrix}{Signature matrix (reccommended odds ratios) of cell-type specificity of genes. Either the object itself or a pathway to an RData file containing an object named "wilcoxon_rank_mat_or" -- generally internal}

\item{DEG_list}{An object with the first column as gene symbols within the bulk dataset (doesn't have to be in signature matrix), second column is the adjusted P-value, and the third the log2FC. Path to a tsv file containing this info is also acceptable.}

\item{control_grep}{Tag in the column name for control (i.e. samples representing downregulated) OR an index of cases.}

\item{max_proportion_change}{Maximum cell-type proportion change -- may be useful if there are many rare cell-types.}

\item{print_plots}{Whether boxplots of the estimated CT proportion for the leave-one-out method of CT deconvolution should be printed.}

\item{plot_names}{If plots are being printed, the prefix of their pdf files.}

\item{theSpecies}{-9 if using a precomputed count matrix from scMappR, human otherwise. removes ensembl symbols if appended.}

\item{make_scale}{convert the lowest odds ratio to 1 and scales accordingly -- strongly not reccomended and will produce warning if used}

\item{FC_coef}{Making scMappR Transformed Values STV's based on fold-change (TRUE) or rank := (-log10(Pval)) (FALSE) rank. After testing, we strongly reccomend to keep true.}

\item{sig_matrix_size}{Number of genes in signature matrix for cell-type deconvolution.}

\item{sig_distort}{Exponential change of odds ratios. Strongly not reccomended and will produce warnings if changed from default.}

\item{drop_unkown_celltype}{Whether or not to remove "unknown" cell-types from the signature matrix.}

\item{toSave}{Allow scMappR to write files in the current directory (T/F)}

\item{ordered_back_all}{output of the g:ProfileR function}

\item{top_bp}{The number of pathways you want to plot}

\item{case_grep.}{Tag in the column name for cases (i.e. samples representing upregulated) OR an index of cases.}
}
\value{
\code{deconvolute_and_contextualize} ScMappR transformed Values for every gene in every cell-type, cell-type composition with, allgenes included, average gene expression of each cell-type usng leave one out approach for each gene, and the processed signature matrix. Optional: boxplots of estimated CT proportions for each gene using a leave-one-out method \cr
}
\description{
This function takes a count matrix, signature matrix, and DEGs before generating STVs for each cell-type.
}
\details{
This function completes the cell-type contextualization in scMappR -- reranking every DEG based on their fold change, likelihood the gene is in each detected cell type, average cell-type proportion, and ratio of cell-type proportion between case and control.
If a gene is upregulated, then it is being controlled by control/case, otherwise it is case/control.
This function expects that the genes within the count file, signature matrix, and DEG_list are have the same logos
It first takes the most "sig_matrix_size" variable genes and completes cell-type deconvolution with it using the DeconRNA-seq package before extracting cell-types with an average of greater than 0.1% of the cell-type proportion's.
In then subsets the signature matrix to only investigate those cell-types for the rest of the analysis. This is because cell-types that are too rare often end with fold-changes between case and control that are unreliable.
Then, it takes a leave-one-out approach to cell-type deconvolution such that estimated cell-type proportions are computed for every gene where they are not in the signature or bulk matrix.
These values are additionally plotted so you can see which genes are contributing the most highly to estimating CT proportions in your bulk sample.
Once these values are computed, cell type means can cell-type ratios between cases and controls for each gene are computed. You can choose to scale your odds ratios such that everything scales to the smallest non-zero value but when using absolute odds ratios it is not reccomented. 
Then, for every gene, scMappR is applied, the example for upreguatled genes is here: val <- cell-preferences * cell-type_proprtion * cell-type_proportion_fold-change * sign*2^abs(gene_DE$log2fc).
Once these values are computed, they are returned. If printplots = true, then this function also returns boplots giving the estimated cell-type proportion with each gene removed.
}
\examples{

data(PBMC_scMappR)
bulk_DE_cors <- PBMC_example$bulk_DE_cors
bulk_normalized <- PBMC_example$bulk_normalized
odds_ratio_in <- PBMC_example$odds_ratio_in
case_grep <- "_female"
control_grep <- "_male"
max_proportion_change <- 10
print_plots <- FALSE
theSpecies <- "human"
norm <- deconvolute_and_contextualize(bulk_normalized, odds_ratio_in, bulk_DE_cors,
                                    case_grep = case_grep, control_grep = control_grep,
                                     max_proportion_change = max_proportion_change,
                                      print_plots = print_plots, 
                                     theSpecies = theSpecies, toSave = F)
                                     
}
