% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gProfiler_STV.R
\name{gProfiler_STV}
\alias{gProfiler_STV}
\title{Pathway enrichment for STV's}
\usage{
gProfiler_STV(STV_matrix, species, background, gene_cut, newGprofiler)
}
\arguments{
\item{STV_matrix}{Matrix of scMappR Transformed Values from the deconvolute_and_contextualize functions.}

\item{species}{Human, mouse, or a charcter that is compatible with gProfileR.}

\item{background}{A list of background genes to test against.}

\item{gene_cut}{The top number of genes in pathway analysis.}

\item{newGprofiler}{Using gProfileR or gprofiler2, (T/F).}
}
\value{
\code{gProfiler_STV} A List of significantly enriched pathways and TFs (correction_method = FDR, hier_sorting = moderate), for every cell-type. \cr
}
\description{
This function runs through each list of scMappR Transformed Values (STV)'s and completes both pathway and Transcription Factor (TF) enrichment.
}
\details{
This function takes a matrix of scMappR_Transformed_Values and a species (human, mouse, or a character directly compatible with g:ProfileR).
Before completing pathway analysis with g:ProfileR. Enriched pathways are stored in a list and returned.
}
\examples{
\donttest{

data(PBMC_example)

bulk_DE_cors <- PBMC_example$bulk_DE_cors
bulk_normalized <- PBMC_example$bulk_normalized
odds_ratio_in <- PBMC_example$odds_ratio_in

case_grep <- "_female"
control_grep <- "_male"
max_proportion_change <- 10
print_plots <- FALSE
theSpecies <- "human"
norm <- deconvolute_and_contextualize(bulk_normalized, odds_ratio_in,
                                      bulk_DE_cors, case_grep = case_grep, 
                                      control_grep = control_grep,
                                       max_proportion_change = max_proportion_change,
                                       print_plots = print_plots,
                                       theSpecies = theSpecies)
background = rownames(bulk_normalized)
STVs <- gProfiler_STV(norm$scMappR_transformed_values, theSpecies,
 background, gene_cut = -9, newGprofiler = FALSE)

 }
 
}
