% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrich_internal.R
\name{pathway_enrich_internal}
\alias{pathway_enrich_internal}
\title{Pathway enrichment for STV's and bulk gene list}
\usage{
pathway_enrich_internal(
  DEGs,
  theSpecies,
  scMappR_vals,
  background_genes,
  output_directory,
  plot_names,
  number_genes,
  newGprofiler,
  toSave = FALSE
)
}
\arguments{
\item{DEGs}{Differentially expressed genes (gene_name, padj, log2fc).}

\item{theSpecies}{Human, mouse, or a charcter that is compatible with gProfileR.}

\item{scMappR_vals}{scMappR Transformed Values of differentially expressed genes.}

\item{background_genes}{A list of background genes to test against.}

\item{output_directory}{Path to the directory where files will be saved.}

\item{plot_names}{Names of output.}

\item{number_genes}{Number of genes to if there are many many DEGs.}

\item{newGprofiler}{Whether to use g:ProfileR or gprofiler2 (T/F).}

\item{toSave}{Allow scMappR to write files in the current directory (T/F).}
}
\value{
\code{pathway_enrich_internal} Plots and pathway enrichment of bulk DE and STVs. \cr
}
\description{
This function completes pathway enrichment of STVs and bulk gene list.
}
\details{
Internal: Pathway analysis od DEGs and STVs for each cell-type. Returns RData objects of differential analysis as well as plots of the top bulk pathways.
It is a wrapper for making barplots, bulk pathway analysis, and gProfiler_STV
}
\examples{
\donttest{

data(PBMC_example)
bulk_DE_cors <- PBMC_example$bulk_DE_cors
bulk_normalized <- PBMC_example$bulk_normalized
odds_ratio_in <- PBMC_example$odds_ratio_in
case_grep <- "_female"
control_grep <- "_male"
max_proportion_change <- 10
print_plots <- FALSE
theSpecies <- "human"
norm <- deconvolute_and_contextualize(bulk_normalized, odds_ratio_in, bulk_DE_cors,
                                       case_grep = case_grep, control_grep = control_grep,
                                        max_proportion_change = max_proportion_change,
                                      print_plots = print_plots, theSpecies = theSpecies)
background = rownames(bulk_normalized)
dir.create("test_path")
pathway_enrich_internal(bulk_DE_cors, "human", norm$scMappR_transformed_values,
                        background, "test_path", "test_figs", toSave = TRUE)

}
}
