% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBP.R
\name{plotBP}
\alias{plotBP}
\title{Plot g:ProfileR Barplot}
\usage{
plotBP(ordered_back_all, top_bp = 10)
}
\arguments{
\item{ordered_back_all}{output of the g:ProfileR function}

\item{top_bp}{The number of pathways you want to plot}
}
\value{
\code{plotBP} A barplot of the number of "top_bp" pathways, ranked by -log10(Pfdr). \cr
}
\description{
Make a barplot of the top biological factors enriched by g:ProfileR
}
\details{
This function takes a g:ProfleR output and prints the top "top_bp" most significantly
enriched P-values before plotting the rank of their P-values#'
}
\examples{
\\notrun {
data("Preoptic_Area")
 POA_generes <- POA_example$POA_generes
 POA_OR_signature <- POA_example$POA_OR_signature
 POA_Rank_signature <- POA_example$POA_Rank_signature
Signature <- as.data.frame(POA_Rank_signature)
rowname <- get_gene_symbol(Signature)
rownames(Signature) <- rowname$rowname
BP <- gProfileR::gprofiler(rowname$rowname, "mmusculus", src_filter = c("GO:BP", "KEGG", "REAC"), max_set_size = 2000, exclude_iea = T)
TF <- gProfileR::gprofiler(rowname$rowname, "mmusculus", src_filter = c("TF"), max_set_size = 5000, exclude_iea = T)
bp <- plotBP(BP)
tf <- make_TF_barplot(TF)

 }
}
