% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dgTMatrix_lists.R
\name{process_dgTMatrix_lists}
\alias{process_dgTMatrix_lists}
\title{Count Matrix To Signature Matrix}
\usage{
process_dgTMatrix_lists(dgTMatrix_list, name, species_name,
  naming_preference, rda_path = "", panglao_set = FALSE,
  haveUMAP = FALSE, saveSCObject = FALSE, internal = FALSE,
  toSave = FALSE)
}
\arguments{
\item{dgTMatrix_list}{A list of matrices in the class of dgTMatrix object -- sparce object -- compatible with Seurat rownames should be of the same species for each.}

\item{name}{The name of the outputted signature matrices, cell-type preferences, and seurat objects if you choose to save them.}

\item{species_name}{Mouse or human symbols, -9 if internal as panglao objects have gene symbol and ensembl strapped together.}

\item{naming_preference}{For cell-type naming, see if cell-types given the inputted tissues are more likely to be named within one of the categories of get_naming_preference_options().}

\item{rda_path}{If saved, directory to where data from scMappR_data is downloaded.}

\item{panglao_set}{If the inputted matrices are from panglao (i.e. if they're internal).}

\item{haveUMAP}{Save the umaps -- only use if the package is downloaded with pip.}

\item{saveSCObject}{Save the seurat object as an RData object (T/F).}

\item{internal}{Was this used as part of the internal processing of panglao datasets (T/F).}

\item{toSave}{Allow scMappR to write files in the current directory (T/F)}
}
\value{
\code{process_from_count} A processed & integrated Seurat object that has been scaled and clustered. It can be returned as an internal object or also stored as an RData object if neccesary. \cr
}
\description{
This function takes a list of count matrices, procces them, calls cell-types, and genreates signature matrices. much of the data is kept.
}
\details{
This function is a one line wrapper to process count matrices into a signature matrix.
It combines process from count, two methods of identifying cell-type identitt (gsva and fisher's test).
Then, it takes the output of cell-type markers and converts it into a signature matrix of P-value ranks and odds-ratios.
Along the way, it saves the Seurat object (if you choose), cell-type identites from gsva (it's own obect), and the signature matrices.
Cell-type marker outputs are also saved in the generes.RData list. Names of the generes objects and the signature matrices are kept.
}
\examples{
\donttest{
data(single_cell_process)
toProcess <- list(example = sm)
tst1 <- process_dgTMatrix_lists(toProcess, "testProcess", -9, 
                                "eye",rda_path = "~/scMappR/data", TRUE)
}
}
