% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMappR_and_pathway_analysis.R
\name{scMappR_and_pathway_analysis}
\alias{scMappR_and_pathway_analysis}
\title{Generate STV, visualize, and enrich.}
\usage{
scMappR_and_pathway_analysis(count_file, signature_matrix, DEG_list,
  case_grep, control_grep, rda_path = "", max_proportion_change = -9,
  print_plots = T, plot_names = "scMappR", theSpecies = "human",
  output_directory = "scMappR_analysis", sig_matrix_size = 3000,
  drop_unkown_celltype = TRUE, internet = TRUE,
  up_and_downregulated = FALSE, gene_label_size = 0.4,
  number_genes = -9, toSave = FALSE)
}
\arguments{
\item{count_file}{Normalized RNA-seq count matrix where rows are gene symbols and columns are individuals. Either the object tself of the path of a TSV file}

\item{signature_matrix}{Signature matrix (reccommended odds ratios) of cell-type specificity of genes. Either the object itself or a pathway to an RData file containing an object named "wilcoxon_rank_mat_or" -- generally internal.}

\item{DEG_list}{An object with the first column as gene symbols within the bulk dataset (doesn't have to be in signature matrix), second column is the adjusted P-value, and the third the log2FC path to a tsv file containing this info is also acceptable.}

\item{case_grep}{Tag in the column name for cases (i.e. samples representing upregulated) OR an index of cases.}

\item{control_grep}{tag in the column name for controls (i.e. samples representing downregulated OR an index of controls.}

\item{max_proportion_change}{Maximum cell-type proportion change -- may be useful if there are many rare cell-types.}

\item{print_plots}{Whether boxplots of the estimated CT proportion for the leave-one-out method of CT deconvolution should be printed. The same name of the plots will be completed for top pathways}

\item{plot_names}{The prefix of plot pdf files.}

\item{theSpecies}{-9 if using a precomputed count matrix from scMappR, human, mouse, or a specied directly compatible with g:Profiler. Removes ensembl symbols if appended}

\item{output_directory}{The name of the directory that will contain off of the analysis.}

\item{sig_matrix_size}{Number of genes in signature matrix for cell-type deconvolution.}

\item{drop_unkown_celltype}{Whether or not to remove "unknown" cell-types from the signature matrix.}

\item{up_and_downregulated}{Whether you are additionally splitting up/downregulated genes (T/F).}

\item{gene_label_size}{The size of the gene label on the plot.}

\item{number_genes}{The number of genes to cut-off for pathway analysis (good with many DEGs)}

\item{toSave}{Allow scMappR to write files in the current directory (T/F)}

\item{WIFI}{Whether you have stable WIFI (T/F).}
}
\value{
\code{scMappR_and_pathway_analysis} A directory with: STVs in RData file, Cell Type proportions (RData file), cell-type proportions leave one out (RData file), heatmap of STVs (all), heatmap of STVs (within signature), heatmap of signature (all), heatmap of signature (overlapping with DEG_list), Pathway enrichment for DEG list(all), RData file and Biological Processes, Pathway enrichment of STVs for each cell-type, RData file and biological processes. \cr
}
\description{
This function generates STVs, visualies them in a heatmap, and completes pathway enrichment of STVs and bulk gene list.
}
\details{
This function generates STVs for every cell-type (see deconvolute_and_contextualize) as well as the relative cell-type proportions (which will be reutrned and pushed through).
Then, it generates heatmaps of all STVs, STVs overlapping with the signature matrix, the signature matrix, the signature matrix overlapping with STVs.
Then, if you have WIFI, it will complete g:ProfilR of the reordered STVs as well as a the ordered list of genes.
This function is a wrapper for deconvolute_and_contextualize, as well as pathway_enrich_internal.
}
\examples{
\\notrun {

load("~/scMappR/data/PBMC_scMappR_and_pathway_analysis_example.rda")
# data(PBMC_scMappR_and_pathway_analysis)
bulk_DE_cors <- PBMC_example$bulk_DE_cors
bulk_normalized <- PBMC_example$bulk_normalized
odds_ratio_in <- PBMC_example$odds_ratio_in
case_grep <- "_female"
control_grep <- "_male"
max_proportion_change <- 10
print_plots <- FALSE
theSpecies <- "human"
toOut <- scMappR_and_pathway_analysis(bulk_normalized, odds_ratio_in, bulk_DE_cors, case_grep = case_grep, control_grep = control_grep, rda_path = "", max_proportion_change = 10, print_plots = T, plot_names = "tst1", theSpecies = "human", output_directory = "tester", sig_matrix_size = 3000, up_and_downregulated = F, internet = F)

}

}
