% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_gene_preferences.R
\name{single_gene_preferences}
\alias{single_gene_preferences}
\title{Single cell-type gene preferences}
\usage{
single_gene_preferences(hg_short, hg_full, study_name, outDir,
  toSave = FALSE)
}
\arguments{
\item{hg_short}{A list with two objects: a "preferences" and a "genesIn". preferences is a list of gene symbols over-represented ine ach cell-type and genesIn were all the inputted genes.}

\item{hg_full}{The same as hg_short but fore very gene in the signature matrix.}

\item{study_name}{name of output table}

\item{toSave}{Allow scMappR to write files in the current directory (T/F)}

\item{outputDir}{directory where table is outputted}
}
\value{
\code{single_gene_preferences} A gene-set enrichment table of individual cell-type enrichment. \cr
}
\description{
Measure enrichment of individual cell-types in a signautre matrix
 
Internal function as part of tissue_scMappR_internal().
This function takes genes preferentially expressed within your gene list, each cell-type
and the background (i.e. all genes within the signature matrix) before completing identifying
the cell-type specific enrichment of the inputted gene list on each cell type.
This function then returns a table describing the cell-type enrichments (p-value and odds ratio)
of each cell-type
}
\examples{
\donttest{

# load in signature matrices
data(Preoptic_Area)
POA_generes <- POA_example$POA_generes
POA_OR_signature <- POA_example$POA_OR_signature
POA_Rank_signature <- POA_example$POA_Rank_signature
Signature <- POA_Rank_signature
genes <- rownames(Signature)[1:60]
heatmap_test <- tissue_scMappR_custom( genes, signature_matrix = Signature,
                                      output_directory =  "scMappR_test", toSave = F)
single_preferences <- heatmap_test$single_celltype_preferences
 }
}
