% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_scMappR_custom.R
\name{tissue_scMappR_custom}
\alias{tissue_scMappR_custom}
\title{Gene List Visualization and Enrichment with Custom Signature Matrix}
\usage{
tissue_scMappR_custom(gene_list, signature_matrix,
  output_directory = "custom_test", toSave = FALSE, gene_cutoff = 1,
  is_pvalue = TRUE)
}
\arguments{
\item{gene_list}{A list of gene symbols, mouse or human.}

\item{signature_matrix}{precomputed signature matrix with matching gene names.}

\item{output_directory}{Directory made containing outputs}

\item{toSave}{Allow scMappR to write files in the current directory (T/F)}

\item{gene_cutoff}{value cutoff (generally rank := log10(Padj)) for a gene to be considered a marker}

\item{is_pvalue}{If signature matrix is p-value before rank is applied (not recommended ) (T/F)}
}
\value{
\code{tissue_scMappR_custom} A list containing the entire signature matrix, the matrix subsetted for your genes, enrichment of each cell-type, and co-enrichment. \cr
}
\description{
This function visualizes signature matrix, cluster's subsetted genes, completes enrichment of individual cell-types and co-enrichment.
}
\details{
This function is roughly the same as tissue_scMappR_internal, however now there is a custom signature matrix.
it generates a heatmap of the signature matrixand your inputted gene list as well as single cell-type and 
co-celltype enrichment.
}
\examples{
\\notrun {

# load in signature matrices
load("data/Preoptic_region_example.rda")
POA_generes <- POA_example$POA_generes
POA_OR_signature <- POA_example$POA_OR_signature
POA_Rank_signature <- POA_example$POA_Rank_signature
# data(Preoptic_region_example)
Signature <- POA_Rank_signature
 rowname <- get_gene_symbol(Signature)
 rownames(Signature) <- rowname$rowname
 genes <- rownames(Signature)[1:200]
 
 # Assuming Signature_matrices_pVal.rda is in you "~/Documents/scMappR/data" directory
 # rda_path1 = "~/Documents/scMappR/data"
 internal <- tissue_scMappR_custom(genes,Signature,output_directory = "scMappR_Test_custom")
 }
}
