% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_scMappR_internal.R
\name{tissue_scMappR_internal}
\alias{tissue_scMappR_internal}
\title{Gene List Visualization and Enrichment (Internal)}
\usage{
tissue_scMappR_internal(
  gene_list,
  species,
  output_directory,
  tissue,
  rda_path = "~/scMappR/data",
  cluster = "Pval",
  genecex = 0.01,
  raw_pval = FALSE,
  toSave = FALSE,
  drop_unkown_celltype = FALSE
)
}
\arguments{
\item{gene_list}{A list of gene symbols, mouse or human.}

\item{species}{"mouse", "human" or "-9" if using a precomputed signature matrix.}

\item{output_directory}{If toSave = TRUE, the name of the output directory that would be built.}

\item{tissue}{Name of the tissue in "get_tissues".}

\item{rda_path}{Path to the .rda file containing all of the signature matrices.}

\item{cluster}{'Pval' or 'OR' depending on if you wantto cluster odds ratios or pvalues of CT preferences.}

\item{genecex}{The size of the gene names of the rows in the heatmap.}

\item{raw_pval}{If the inputted signature matrix are raw (untransformed) p-values -- recommended to generate rank first (T/F).}

\item{toSave}{Allow scMappR to write files in the current directory (T/F).}

\item{drop_unkown_celltype}{Whether or not to remove "unknown" cell-types from the signature matrix (T/F).}
}
\value{
\code{tissue_scMappR_internal} A list containing the entire signature matrix, the matrix subsetted for your genes, enrichment of each cell-type, and co-enrichment. \cr
}
\description{
This function loops through every signature matrix in a particular tissue and generates heatmaps, cell-type preferences, and co-enrichment.
}
\details{
This function takes a list of genes and a tissue that is contained in current signature matrices before and generating heatmaps of cell-type preferences.
It then completes cell-type enrichment of each individual cell-type, then, if more than two cell-types are signficiantly enriched, co-enrichemnt 
of those enriched cell-types is then computed.
}
\examples{
\donttest{

data(POA_example)
POA_generes <- POA_example$POA_generes
POA_OR_signature <- POA_example$POA_OR_signature
POA_Rank_signature <- POA_example$POA_Rank_signature
Signature <- POA_Rank_signature
rowname <- get_gene_symbol(Signature)
rownames(Signature) <- rowname$rowname
genes <- rownames(Signature)[1:200]
rda_path1 = "~/scMappR/data"
internal <- tissue_scMappR_internal(genes,"mouse", output_directory = "scMappR_Test",
                                    tissue = "hypothalamus",rda_path = rda_path1)

 }
}
