% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toNum.R
\name{toNum}
\alias{toNum}
\title{To Numeric.}
\usage{
toNum(x)
}
\arguments{
\item{x}{A character, factor, or numeric vector.}
}
\value{
\code{toNum} Returns a numeric vector. \cr
}
\description{
This function checks if your vector is not a character and if it is, then converts it to a numeric.
}
\examples{

 
 # vector of factors
fact <- factor(c("1", "2", "3", "4"))
# convert to numeric
num <- toNum(fact)
 
}
